Port plastic shader to the new system.
Move some glass utility functions to vector.h
