+++ src/renderer/bsdf/glass.c
@@ -21,28 +21,6 @@ struct glassBsdf {
 	struct textureNode *color;
 };
 
-static inline bool refract(const struct vector *in, const struct vector normal, float niOverNt, struct vector *refracted) {
-	const struct vector uv = vecNormalize(*in);
-	const float dt = vecDot(uv, normal);
-	const float discriminant = 1.0f - niOverNt * niOverNt * (1.0f - dt * dt);
-	if (discriminant > 0.0f) {
-		const struct vector A = vecScale(normal, dt);
-		const struct vector B = vecSub(uv, A);
-		const struct vector C = vecScale(B, niOverNt);
-		const struct vector D = vecScale(normal, sqrtf(discriminant));
-		*refracted = vecSub(C, D);
-		return true;
-	} else {
-		return false;
-	}
-}
-
-static inline float schlick(float cosine, float IOR) {
-	float r0 = (1.0f - IOR) / (1.0f + IOR);
-	r0 = r0 * r0;
-	return r0 + (1.0f - r0) * powf((1.0f - cosine), 5.0f);
-}
-
 struct bsdfSample glass_sample(const struct bsdf *bsdf, sampler *sampler, const struct hitRecord *record, const struct vector *in) {
 	struct glassBsdf *glassBsdf = (struct glassBsdf*)bsdf;
